// shared/src/interfaces/jobInterface.ts
/**
 * @fileoverview Defines the IJob interface for job data.
 * @version 1.0.0
 * @module jobInterface
 */

import { Optional } from 'sequelize';

/**
 * Interface representing the attributes of a job.
 */
export interface IJob {
  id?: number;
  title: string;
  description: string;
  responsibility: string;
  requirements: string[];
  benefits?: string[];
  location: string;
  type: 'Full-time' | 'Part-time' | 'Contract' | 'Freelance' | 'Internship';
  vacancy?: number;
  deadline: Date;
  tags: string[];
  recruiterId: number;
  views?: number;
  isActive?: boolean;
  createdAt?: Date;
  updatedAt?: Date;
}

/**
 * Interface for creating a new job (excluding autogenerated fields).
 */
export type IJobCreation = Optional<
  IJob,
  | 'id'
  | 'benefits'
  | 'vacancy'
  | 'views'
  | 'isActive'
  | 'createdAt'
  | 'updatedAt'
>;